Fixed account unique constraint issue

+++ accounts/src/main/java/sg/darren/microservices/accounts/model/entity/Customer.java
@@ -18,7 +18,7 @@ public class Customer {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "customer_id")
-    private int customerId;
+    private Long customerId;
 
     private String name;
