+++ accounts/src/main/java/sg/darren/microservices/accounts/repository/AccountRepository.java
@@ -7,6 +7,6 @@ import sg.darren.microservices.accounts.model.entity.Account;
 @Repository
 public interface AccountRepository extends JpaRepository<Account, Long> {
 
-    Account findByCustomerId(int customerId);
+    Account findByCustomerId(Long customerId);
 
 }
